working

+++ src/main/java/com/odian/moviesearch/api/model/PagedResponseDTO.java
@@ -3,7 +3,7 @@ package com.odian.moviesearch.api.model;
 import java.util.List;
 
 public record PagedResponseDTO<T>(
-    Integer totalItems,
+    Long totalItems,
     Integer totalPages,
     Integer currentPage,
     Integer pageSize,
